Change appBar to a component
